Deletes completed

+++ Backend/Proyecto/src/main/java/com/proyectoJuegoCalabozos/Proyecto/Controller/MonsterController.java
@@ -3,7 +3,9 @@ package com.proyectoJuegoCalabozos.Proyecto.Controller;
 import java.util.List;
 
 import com.proyectoJuegoCalabozos.Proyecto.Model.Monster;
+import com.proyectoJuegoCalabozos.Proyecto.Model.MonstersEsp;
 import com.proyectoJuegoCalabozos.Proyecto.Repository.MonsterRepository;
+import com.proyectoJuegoCalabozos.Proyecto.Repository.MonstersEspRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,6 +23,9 @@ public class MonsterController {
     @Autowired
     private MonsterRepository monsterRepository;
 
+    @Autowired
+    private MonstersEspRepository monstersRepository;
+
     @GetMapping("/all")
     public String allMonsters(Model model) {
         List<Monster> monsters = monsterRepository.findAll();
@@ -31,6 +36,8 @@ public class MonsterController {
     @GetMapping("/create")
     public String createMonster(Model model) {
         Monster monster = new Monster("",0);
+        List<MonstersEsp> monsters = monstersRepository.findAll();
+        model.addAttribute("monsters", monsters);
         model.addAttribute("monster", monster);
         return "Monster-templates/Monster-create";
     }
@@ -56,7 +63,15 @@ public class MonsterController {
 
     @GetMapping("/delete/{id}")
     public String deleteMonster(Model model, @PathVariable Long id) {
+        Monster m = monsterRepository.findById(id).get();
+        
+        if(m.getMonsterEsp()==null||m.getRoom()==null)
         monsterRepository.deleteById(id);
+        else {
+            m.unlink(m.getRoom(), m.getMonsterEsp());
+            monsterRepository.deleteById(id);
+        }
+        
         return "redirect:/monster/all";
     }
